requirements.txt
# Hamlet QA failure-analysis harness
torch>=2.1.0
transformers>=4.40.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
vllm>=0.4.0
numpy>=1.24.0
tqdm>=4.65.0
